Refactor project structure to match FinniversKit
