--- FilterKit.xcworkspace/contents.xcworkspacedata
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Workspace
-   version = "1.0">
-   <FileRef
-      location = "group:Demo/Demo.xcodeproj">
-   </FileRef>
-   <FileRef
-      location = "group:FilterKit/FilterKit.xcodeproj">
-   </FileRef>
-</Workspace>
