--- FilterKit/Sources/FilterKit.h
@@ -1,11 +1,0 @@
-#import <UIKit/UIKit.h>
-
-//! Project version number for FilterKit.
-FOUNDATION_EXPORT double FilterKitVersionNumber;
-
-//! Project version string for FilterKit.
-FOUNDATION_EXPORT const unsigned char FilterKitVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <FilterKit/PublicHeader.h>
-
-
