+++ Sources/Bottomsheet/BottomsheetTransitioningAnimator.swift
@@ -4,27 +4,27 @@
 
 import UIKit
 
-final public class BottomsheetTransitioningAnimator: NSObject, UIViewControllerAnimatedTransitioning {
+public final class BottomsheetTransitioningAnimator: NSObject, UIViewControllerAnimatedTransitioning {
     public enum TransitionType {
         case presentation, dismissal
     }
-    
+
     public var transitionType: TransitionType?
 
     public func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
         guard let transitionType = transitionType else {
             return 0.0
         }
-        
+
         return transitionType == .presentation ? 0.4 : 0.4
     }
-    
+
     public func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
         guard let transitionType = transitionType else {
             transitionContext.completeTransition(false)
             return
         }
-        
+
         switch transitionType {
         case .presentation:
             animatePresentationTransition(using: transitionContext)
@@ -40,34 +40,33 @@ private extension BottomsheetTransitioningAnimator {
             transitionContext.completeTransition(false)
             return
         }
-        
+
         let finalFrame = transitionContext.finalFrame(for: presentedViewController)
         presentedViewController.view.frame = transitionContext.finalFrame(for: presentedViewController)
-        
+
         transitionContext.containerView.insertSubview(presentedViewController.view, aboveSubview: presentingViewController.view)
-        
+
         let animationDuration = transitionDuration(using: transitionContext)
         let offScreenTransform = CGAffineTransform(translationX: 0, y: transitionContext.containerView.frame.maxY - finalFrame.origin.y)
         presentedViewController.view.transform = offScreenTransform
-        
-        
+
         UIView.animate(withDuration: animationDuration, delay: 0.0, usingSpringWithDamping: 0.7, initialSpringVelocity: 1.0, options: [.curveEaseInOut], animations: {
             presentedViewController.view.transform = .identity
         }) { _ in
             transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
         }
     }
-    
+
     func animateDismissalTransition(using transitionContext: UIViewControllerContextTransitioning) {
         guard let dismissingViewController = transitionContext.viewController(forKey: .from) else {
             transitionContext.completeTransition(false)
             return
         }
-        
+
         let animationDuration = transitionDuration(using: transitionContext)
         let finalFrame = transitionContext.finalFrame(for: dismissingViewController)
         let offScreenTransform = CGAffineTransform(translationX: 0, y: transitionContext.containerView.frame.maxY - finalFrame.origin.y)
-        
+
         UIView.animate(withDuration: animationDuration, animations: {
             dismissingViewController.view.transform = offScreenTransform
         }) { _ in
@@ -76,8 +75,7 @@ private extension BottomsheetTransitioningAnimator {
     }
 }
 
-
-//func animatePresentationTransition(using transitionContext: UIViewControllerContextTransitioning) {
+// func animatePresentationTransition(using transitionContext: UIViewControllerContextTransitioning) {
 //    guard let presentedViewController = transitionContext.viewController(forKey: .to), let presentingViewController = transitionContext.viewController(forKey: .from) else {
 //        transitionContext.completeTransition(false)
 //        return
@@ -96,9 +94,9 @@ private extension BottomsheetTransitioningAnimator {
 //    }) { _ in
 //        transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
 //    }
-//}
+// }
 //
-//func animateDismissalTransition(using transitionContext: UIViewControllerContextTransitioning) {
+// func animateDismissalTransition(using transitionContext: UIViewControllerContextTransitioning) {
 //    guard let dismissingView = transitionContext.view(forKey: .from) else {
 //        transitionContext.completeTransition(false)
 //        return
@@ -112,4 +110,4 @@ private extension BottomsheetTransitioningAnimator {
 //    }) { _ in
 //        transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
 //    }
-//}
+// }
