+++ Sources/FilterKit.h
@@ -1,0 +1,8 @@
+//
+//  Copyright © FINN.no AS, Inc. All rights reserved.
+//
+
+@import UIKit;
+
+FOUNDATION_EXPORT double FilterKitVersionNumber;
+FOUNDATION_EXPORT const unsigned char FilterKitVersionString[];
