+++ UnitTests/UnitTests.swift
@@ -1,0 +1,8 @@
+//
+//  Copyright © FINN.no AS, Inc. All rights reserved.
+//
+
+import XCTest
+
+class UnitTests: XCTestCase {
+}
